Classes to build from the API have been slimmed to Artist, Venue, and Event(Show in DB). Each class has itself been trimmed to having only the attributes needed to allow for functionality -- while also attaining parity with the DB schema.
Note: The previous code has been placed in LMNFlask\utils\sk_api_v_1\. We can easily access any portion of it as needed, while allowing for the new code to be placed in the exact same class paths, preserving import integrity.

Also, Macs are the Negan to my Rick, may the upcoming finale be without spoilers and bring a return to the days of a not-totally-lame story telling.